FLEX-4697 ~ Solves Quartz scheduler resource leaks.
- solves worker-threads not stopping thus leaking memory
- solves HikariDB-pool-threads not stopping and pool not closing thus
leaking memory
