+++ osgp/platform/osgp-adapter-ws-admin/src/main/java/org/opensmartgridplatform/adapter/ws/admin/application/config/PersistenceConfig.java
@@ -7,17 +7,16 @@
  */
 package org.opensmartgridplatform.adapter.ws.admin.application.config;
 
-import javax.annotation.PreDestroy;
+import javax.sql.DataSource;
 
+import org.opensmartgridplatform.domain.core.repositories.DeviceRepository;
+import org.opensmartgridplatform.shared.application.config.AbstractPersistenceConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-
-import org.opensmartgridplatform.domain.core.repositories.DeviceRepository;
-import org.opensmartgridplatform.shared.application.config.AbstractPersistenceConfig;
 
 @EnableJpaRepositories(entityManagerFactoryRef = "entityManagerFactory", basePackageClasses = {
         DeviceRepository.class })
@@ -26,6 +25,12 @@ import org.opensmartgridplatform.shared.application.config.AbstractPersistenceCo
 @PropertySource(value = "file:${osgp/Global/config}", ignoreResourceNotFound = true)
 @PropertySource(value = "file:${osgp/AdapterWsAdmin/config}", ignoreResourceNotFound = true)
 public class PersistenceConfig extends AbstractPersistenceConfig {
+
+    @Override
+    @Bean(destroyMethod = "close")
+    public DataSource getDataSource() {
+        return super.getDataSource();
+    }
 
     @Override
     @Bean
@@ -37,12 +42,6 @@ public class PersistenceConfig extends AbstractPersistenceConfig {
     @Bean
     public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
         return super.entityManagerFactory("OSGP_WS_ADAPTER_ADMIN");
-    }
-
-    @Override
-    @PreDestroy
-    public void destroyDataSource() {
-        super.destroyDataSource();
     }
 
 }
