+++ osgp/platform/osgp-adapter-ws-distributionautomation/src/main/resources/osgp-adapter-ws-distributionautomation.properties
@@ -112,14 +112,14 @@ jms.distributionautomation.logging.queue=osgp.logging.ws
 # SCHEDULING CONFIG
 # =========================================================
 
+quartz.scheduler.thread.count=2
+
 ## Production (every day at 10:00 AM) ->  0 0 10 * * ?
 ## Test (every minute)                  ->  0 0/1 * * * ?
 distributionautomation.scheduling.job.cleanup.response.data.cron.expression=0 0/1 * * * ?
-distributionautomation.scheduling.job.cleanup.response.data.thread.count=1
 distributionautomation.scheduling.job.cleanup.response.data.retention.time.in.days=7
 
 distributionautomation.scheduling.job.resend.notification.cron.expression=0 0/1 * * * ?
-distributionautomation.scheduling.job.resend.notification.thread.count=1
 distributionautomation.scheduling.job.resend.notification.pool.size=1
 distributionautomation.scheduling.job.resend.notification.resend.threshold.in.minutes=2
 distributionautomation.scheduling.job.resend.notification.multiplier=2
