+++ osgp/platform/osgp-adapter-ws-microgrids/src/main/java/org/opensmartgridplatform/adapter/ws/microgrids/application/config/PersistenceConfigWs.java
@@ -7,7 +7,10 @@
  */
 package org.opensmartgridplatform.adapter.ws.microgrids.application.config;
 
+import javax.sql.DataSource;
+
 import org.flywaydb.core.Flyway;
+import org.opensmartgridplatform.shared.application.config.AbstractPersistenceConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
@@ -15,8 +18,6 @@ import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-
-import org.opensmartgridplatform.shared.application.config.AbstractPersistenceConfig;
 
 @EnableJpaRepositories(transactionManagerRef = "transactionManager", entityManagerFactoryRef = "wsEntityManagerFactory", basePackageClasses = {
         org.opensmartgridplatform.adapter.ws.domain.repositories.ResponseDataRepository.class })
@@ -28,6 +29,11 @@ public class PersistenceConfigWs extends AbstractPersistenceConfig {
 
     public PersistenceConfigWs() {
         // Empty default constructor
+    }
+
+    @Bean(destroyMethod = "close")
+    public DataSource dataSource() {
+        return super.getDataSource();
     }
 
     @Override
