+++ osgp/platform/osgp-adapter-ws-microgrids/src/main/resources/osgp-adapter-ws-microgrids.properties
@@ -120,14 +120,14 @@ jms.microgrids.logging.time.to.live=3600000
 # SCHEDULING CONFIG
 # =========================================================
 
+quartz.scheduler.thread.count=2
+
 ## Production (every day at 10:00 AM) ->  0 0 10 * * ?
 ## Test (every minute)                  ->  0 0/1 * * * ?
 microgrids.scheduling.job.cleanup.response.data.cron.expression=0 0/1 * * * ?
-microgrids.scheduling.job.cleanup.response.data.thread.count=1
 microgrids.scheduling.job.cleanup.response.data.retention.time.in.days=7
 
 microgrids.scheduling.job.resend.notification.cron.expression=0 0/1 * * * ?
-microgrids.scheduling.job.resend.notification.thread.count=1
 microgrids.scheduling.job.resend.notification.pool.size=1
 microgrids.scheduling.job.resend.notification.resend.threshold.in.minutes=2
 microgrids.scheduling.job.resend.notification.multiplier=2
