+++ osgp/platform/osgp-adapter-ws-publiclighting/src/main/resources/osgp-adapter-ws-publiclighting.properties
@@ -53,7 +53,6 @@ web.service.notification.enabled=true
 web.service.notification.application.name=
 
 publiclighting.scheduling.job.resend.notification.cron.expression=0 0/1 * * * ?
-publiclighting.scheduling.job.resend.notification.thread.count=1
 #Disable notification re-sending for now.
 publiclighting.scheduling.job.resend.notification.maximum=0
 
@@ -161,8 +160,9 @@ jms.publiclighting.logging.use.exponential.back.off=true
 # SCHEDULING CONFIG
 # =========================================================
 
+quartz.scheduler.thread.count=2
+
 ## Production (every day at 10:00 AM) ->  0 0 10 * * ?
 ## Test (every minute)                  ->  0 0/1 * * * ?
 publiclighting.scheduling.job.cleanup.response.data.cron.expression=0 0/1 * * * ?
-publiclighting.scheduling.job.cleanup.response.data.thread.count=1
 publiclighting.scheduling.job.cleanup.response.data.retention.time.in.days=1
