+++ osgp/platform/osgp-adapter-ws-smartmetering/src/main/resources/osgp-adapter-ws-smartmetering.properties
@@ -117,14 +117,14 @@ sync.notification.delay=5000
 # SCHEDULING CONFIG
 # =========================================================
 
+quartz.scheduler.thread.count=2
+
 ## Production (every day at 10:00 AM) ->  0 0 10 * * ?
 ## Test (every minute)                  ->  0 0/1 * * * ?
 smartmetering.scheduling.job.cleanup.response.data.cron.expression=0 0/1 * * * ?
-smartmetering.scheduling.job.cleanup.response.data.thread.count=1
 smartmetering.scheduling.job.cleanup.response.data.retention.time.in.days=7
 
 smartmetering.scheduling.job.resend.notification.cron.expression=0 0/1 * * * ?
-smartmetering.scheduling.job.resend.notification.thread.count=1
 smartmetering.scheduling.job.resend.notification.pool.size=1
 smartmetering.scheduling.job.resend.notification.resend.threshold.in.minutes=2
 smartmetering.scheduling.job.resend.notification.multiplier=2
