+++ osgp/platform/osgp-adapter-ws-tariffswitching/src/main/resources/osgp-adapter-ws-tariffswitching.properties
@@ -47,7 +47,6 @@ web.service.notification.enabled=true
 web.service.notification.application.name=
 
 tariffswitching.scheduling.job.resend.notification.cron.expression=0 0/1 * * * ?
-tariffswitching.scheduling.job.resend.notification.thread.count=1
 #Disable notification re-sending for now.
 tariffswitching.scheduling.job.resend.notification.maximum=0
 
@@ -150,8 +149,9 @@ jms.tariffswitching.logging.use.exponential.back.off=true
 # SCHEDULING CONFIG
 # =========================================================
 
+quartz.scheduler.thread.count=2
+
 ## Production (every day at 10:00 AM) ->  0 0 10 * * ?
 ## Test (every minute)                  ->  0 0/1 * * * ?
 tariffswitching.scheduling.job.cleanup.response.data.cron.expression=0 0/1 * * * ?
-tariffswitching.scheduling.job.cleanup.response.data.thread.count=1
 tariffswitching.scheduling.job.cleanup.response.data.retention.time.in.days=1
