+++ osgp/shared/shared/src/main/java/org/opensmartgridplatform/shared/application/config/AbstractPersistenceConfig.java
@@ -14,17 +14,16 @@ import javax.sql.DataSource;
 import org.flywaydb.core.Flyway;
 import org.flywaydb.core.api.MigrationVersion;
 import org.hibernate.ejb.HibernatePersistence;
+import org.opensmartgridplatform.shared.infra.db.DefaultConnectionPoolFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 
-import org.opensmartgridplatform.shared.infra.db.DefaultConnectionPoolFactory;
 import com.zaxxer.hikari.HikariDataSource;
 
 /**
  * This class provides the basic components used for persistence.
  */
-
 public abstract class AbstractPersistenceConfig extends AbstractConfig {
 
     private static final String REGEX_COMMA_WITH_OPTIONAL_WHITESPACE = "\\s*+,\\s*+";
